# api-spec.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://127.0.0.1:8080","description":"Generated server url"}],"tags":[{"name":"Контакты","description":"Методы для работы с контактной информацией клиентов"},{"name":"Клиенты","description":"Методы для работы с клиентами"}],"paths":{"/clients":{"get":{"tags":["Клиенты"],"summary":"Получение постраничного списка клиентов.","operationId":"getClients","parameters":[{"name":"from","in":"query","required":false,"schema":{"minimum":0,"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClientDto"}}}}}}},"post":{"tags":["Клиенты"],"summary":"Добавление нового клиента.","operationId":"addClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewClientDto"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}}}}},"/clients/{clientId}/contacts":{"get":{"tags":["Контакты"],"summary":"Получение контактов клиента.","operationId":"getContacts","parameters":[{"name":"clientId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"type","in":"query","required":false,"schema":{"type":"string","enum":["EMAIL","PHONE"]}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ContactDto"}}}}}}},"post":{"tags":["Контакты"],"summary":"Добавление контакта клиенту.","operationId":"addContact","parameters":[{"name":"clientId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewContactDto"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ContactDto"}}}}}}},"/clients/{clientId}":{"get":{"tags":["Клиенты"],"summary":"Получение данных клиента по его идентификатору","operationId":"getClient","parameters":[{"name":"clientId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}}}}}},"components":{"schemas":{"ApiError":{"type":"object","properties":{"status":{"type":"string","description":"Код статуса HTTP-ответа"},"reason":{"type":"string","description":"Общее описание причины ошибки."},"message":{"type":"string","description":"Сообщение об ошибке."},"timestamp":{"type":"string","description":"Дата и время когда произошла ошибка (в формате \"yyyy-MM-dd HH:mm:ss\")"},"errors":{"type":"array","description":"Список стектрейсов или описания ошибок.","items":{"type":"string","description":"Список стектрейсов или описания ошибок."}}},"description":"Описание ошибки"},"NewClientDto":{"required":["name"],"type":"object","properties":{"name":{"maxLength":100,"minLength":0,"type":"string","description":"ФИО"}},"description":"Информация о новом клиенте."},"ClientDto":{"type":"object","properties":{"id":{"type":"integer","description":"Идентификатор","format":"int64"},"name":{"type":"string","description":"ФИО"}},"description":"Информация о полученном клиенте."},"NewContactDto":{"required":["description","type"],"type":"object","properties":{"type":{"type":"string","description":"Тип контактной информации","enum":["EMAIL","PHONE"]},"description":{"maxLength":100,"minLength":0,"type":"string","description":"Значение"}},"description":"Информация о новой контактной информации"},"ContactDto":{"type":"object","properties":{"id":{"type":"integer","description":"Идентификатор","format":"int64"},"type":{"type":"string","description":"Тип","enum":["EMAIL","PHONE"]},"description":{"type":"string","description":"Значение"}},"description":"Информация о полученной контактной информации"}}}}